+++ src/main/java/de/glmtk/smoothing/estimator/fraction/MaximumLikelihoodEstimator.java
@@ -7,7 +7,7 @@ public class MaximumLikelihoodEstimator extends FractionEstimator {
     @Override
     protected double calcNumerator(NGram sequence, NGram history, int recDepth) {
         NGram fullSequence = getFullSequence(sequence, history);
-        int fullSequenceCount = countCache.getAbsolute(fullSequence);
+        long fullSequenceCount = countCache.getAbsolute(fullSequence);
         logDebug(recDepth, "fullSequence = {} ({})", fullSequence,
                 fullSequenceCount);
         return fullSequenceCount;
@@ -15,9 +15,9 @@ public class MaximumLikelihoodEstimator extends FractionEstimator {
 
     @Override
     protected double
-        calcDenominator(NGram sequence, NGram history, int recDepth) {
+    calcDenominator(NGram sequence, NGram history, int recDepth) {
         NGram fullHistory = getFullHistory(sequence, history);
-        int fullHistoryCount = countCache.getAbsolute(fullHistory);
+        long fullHistoryCount = countCache.getAbsolute(fullHistory);
         logDebug(recDepth, "fullHistory = {} ({})", fullHistory,
                 fullHistoryCount);
         return fullHistoryCount;
